Refactored writer-test.js to improve OCP and SRP based on feedback.  Added string-writer-test.js and file-writer-test.js files
git-svn-id: https://agile.cs.uh.edu/svn/f17_softwaredesign/berejnoi_biddle@2995 5f3e6de7-e054-8643-865b-9ca9e3633ffd

+++ assign3/test/writer-test.js
@@ -6,20 +6,21 @@ const FileWriter = require('../src/file-writer');
 
 const fs = require('fs-extra');
 
-describe('writer tests:', function(){
+const createTests = function(WriterClass) {
+	describe('writer tests:', function(){
 
-	it('canary test', function(){
-		expect(true).to.be.true;
-	});
+		const TEST_OUTPUT_FILE = './test.txt';
 
-	const TEST_OUTPUT_FILE = './test.txt';
+		before(() => writer = new WriterClass(TEST_OUTPUT_FILE));
 
-	writers = [ new FileWriter(TEST_OUTPUT_FILE),
-	 			new StringWriter()]; //Venkat: This file is violating SRP and OCP.
+		if (WriterClass == FileWriter){
+			after(() => fs.unlinkSync(TEST_OUTPUT_FILE));
+		}
 
-	after(() => fs.unlinkSync(TEST_OUTPUT_FILE));
+		it('canary test', function(){
+			expect(true).to.be.true;
+		});
 
-	writers.forEach((writer) =>{
 		it('should open and write', function(){
 			return writer.write("some string")
 				.then(() => {
@@ -56,6 +57,6 @@ describe('writer tests:', function(){
 				});
 		});
 	});
+}
 
-	
-});
+module.exports = createTests;
